Simulation analysis on Diphtheria
This is a secondary data gotten from Nigeria Centre for Disease Control. 
The code show the data visualization for different variables. Also multiple regression model is used to show the independent variables (Lab confirmed, epidemiologically linked and clinical compatible) can contribute to the number of deaths.
The model predicts that Mauritania and Niger have a higher probability of a high Case Fatality Ratio (CFR) compared to other countries and may require interventions to reduce the mortality rates.